Initialization/product seed

+++ database/migrations/2022_10_19_062615_create_products_table.php
@@ -18,7 +18,7 @@ class CreateProductsTable extends Migration
             $table->string('product_image');
             $table->string('product_name');
             $table->string('product_description');
-            $table->integer('product_price');
+            $table->bigInteger('product_price');
             $table->unsignedBigInteger('category_id');
             $table->timestamps();
